requirements.txt
vllm>=0.6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
oss2>=2.18.0
